Move unit converter helper to composition package
