+++ platform/autos/src/main/java/br/com/orcinus/orca/platform/autos/kit/input/text/composition/Units.java
@@ -13,14 +13,14 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.platform.autos.kit;
+package br.com.orcinus.orca.platform.autos.kit.input.text.composition;
 
 import android.content.Context;
 import android.util.TypedValue;
 import androidx.annotation.NonNull;
 
 /** Utilities for performing operations with units. */
-public class Units {
+class Units {
   private Units() {}
 
   /**
@@ -29,7 +29,7 @@ public class Units {
    * @param context {@link Context} through which the conversion will take place.
    * @param dp Amount in DPs to be converted into pixels.
    */
-  public static int dp(@NonNull Context context, float dp) {
+  static int dp(@NonNull Context context, float dp) {
     return (int)
         TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP, dp, context.getResources().getDisplayMetrics());
